ALCACHOFA: Split up OpenGL renderer
